ai-engine/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic>=2.7.0
chromadb
pypdf
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-chroma>=0.1.4
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-neo4j>=0.0.1
langchain-experimental>=0.3.0,<0.4.0
edge-tts
mutagen
httpx
